drm/atomic: Boost DDR bus when committing a new frame
In order to reduce jank, request and DDR bus boost whenever a new
frame is ready to be rendered to the display. The boosts should be
sufficient enough to render 60 FPS without any dropped frames when there
is no significant external source of load.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: ArbitraryFox <chrisung1992@gmail.com>

Remove the cpu_input_boost option since we only boost for app launch and screen on